requirements.txt
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
filelock==3.20.0
fsspec==2025.9.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
maturin==1.9.6
mpmath==1.3.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pytokens==0.2.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.9.0
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
yara-python==4.5.1